requirements.txt
nicegui>=2.0.0
google-generativeai>=0.8.0
requests>=2.32.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
gunicorn==21.2.0
